Batch renderer functional.

+++ Hazel/src/Hazel/Rendering/Batch/Vertex.h
@@ -10,12 +10,12 @@ namespace Hazel
 			{"a_Position", ShaderDataType::Float3},
 			{"a_Color", ShaderDataType::Float4},
 			{"a_TextureCoordinate", ShaderDataType::Float2},
-			{"a_TextureIndex", ShaderDataType::Integer}
+			{"a_TextureIndex", ShaderDataType::Float}
 		};
 
 		glm::vec3 Position{0.0f};
 		glm::vec4 Color{1.0f};
 		glm::vec2 TextureCoordinate{0.0f};
-		int TextureIndex = 0;
+		float TextureIndex = 0.0f;
 	};
 }
